remove waiting for unitChanged after controlUnit
Previously I assumed that we can wait for unitChanged event after sending
controlUnit to detect whether the gateway is responding or not. Turns out
this doesn't work if the sent controlUnit doesn't actually change the
state of the light (in that case unitChanged rightly isn't sent).

Removing this logic means that HomeKit may get out of sync if the gateway
is not responding but that might be ok since things are broken anyway.