add support for sites
The API bindings have been extended to support multiple network sessions
on a single WebSocket connection (one wire per network session).

The user can now specify whether the credentials provided in the config
are for a network or a user account.

For a user account, all networks in all sites belonging to that user
are scanned for units and added to HomeKit.