replace controlUnit timeout with unitChanged delay
When the state of a device changes, the cloud sends a unitChanged event. This happens also when
the change was caused by sending a controlUnit message.

The plugin handles the event by passing the new state to HomeKit.

This leads to issues when, for example, the user is sliding the brighness bar in the Home App.
The values from the unitChanged events are moving the slider while the user is using it, creating
weird UI glitches.

To ignore those events, a controlUnit timeout was originally implemented. It caused the first
unitChanged event received after sending controlUnit to be ignored.

This didn't work well and ignoring events wasn't a good idea to begin with. It has now been replaced
with a delay built into the unitChanged handler. When an event is received, the HomeKit characteristics
are updated after a short delay. If more events are received within that time, the previous updates
are ignored. This ensures that HomeKit is not flooded with updates but at the same time it always
eventually reflects the real latest state of the device.