add support for "Vertical" control type to luminaries
This adds support for controlling the direction of the light (up/down) of luminaries
that have that feature, like Occhio's Sento or Mitto largo lamps.

The direction is exposed by the Cloud API as "Vertical" control type, expressed as
a float between 0.0 (uplight only) and 1.0 (downlight only).

Since HomeKit doesn't support such a characteristic yet, the plugin adds a second
Lightbulb service to the accessory and then uses its Brightness characteristic to
control the direction.

This is shown in the Home app as a single tile with two sliders when opened.
If desired, there's an option to split the services into two separate tiles.

The on/off state of the second Lightbulb service mimics the state main service (they're
either both on or both off).